refactor: delete ReplicaStats; inline display fields into cli consumers
`ReplicaStats` was a 3-field struct in `dag/metrics/snapshot.rs` whose
only reason for existing was display: `#[tabled(inline)]` on cli's
`ReplicaRow` and `#[serde(flatten)]` on cli's `ReplicaReport`. The
domain crate carried cli-specific derive attributes — out of place.

Delete the struct, its `replica_stats()` accessor, and the
`fmt_latency_ms` helper; inline `p50_latency_ms` / `p90_latency_ms` /
`leader_timeouts` as direct fields on each consumer:
- `cli/src/terminal/table.rs::ReplicaRow` — direct `#[tabled]`
  attributes on the three fields; local `fmt_latency_ms` helper for
  `display_with`.
- `cli/src/report.rs::ReplicaReport` — direct fields;
  `Option<f64>` serialises natively, no display helper needed.

Values come from the new `snapshot.latency_percentile_ms(p)` and a
direct `snapshot.metric(LEADER_TIMEOUT_TOTAL, &[]) as u64` read.

Refs #66.
