+++ crates/cli/src/terminal/table.rs
@@ -3,7 +3,7 @@
 
 use dag::{
     authority::Authority,
-    metrics::{MetricsSnapshot, Outcome, ReplicaStats, RunResult},
+    metrics::{LEADER_TIMEOUT_TOTAL, MetricsSnapshot, Outcome, RunResult},
 };
 use simulator::SimulationConfig;
 use tabled::{Table, Tabled, settings::Style};
@@ -55,8 +55,12 @@ pub struct ReplicaRow {
     committed_leaders: usize,
     #[tabled(rename = "commits/s")]
     commits_per_sec: String,
-    #[tabled(inline)]
-    stats: ReplicaStats,
+    #[tabled(rename = "p50 latency", display_with = "fmt_latency_ms")]
+    p50_latency_ms: Option<f64>,
+    #[tabled(rename = "p90 latency", display_with = "fmt_latency_ms")]
+    p90_latency_ms: Option<f64>,
+    #[tabled(rename = "leader timeouts")]
+    leader_timeouts: u64,
 }
 
 impl ReplicaRow {
@@ -78,7 +82,9 @@ impl ReplicaRow {
             replica: authority,
             committed_leaders,
             commits_per_sec,
-            stats: metrics.replica_stats(),
+            p50_latency_ms: metrics.latency_percentile_ms(0.5),
+            p90_latency_ms: metrics.latency_percentile_ms(0.9),
+            leader_timeouts: metrics.metric(LEADER_TIMEOUT_TOTAL, &[]) as u64,
         }
     }
 
@@ -138,5 +144,12 @@ impl SuiteRow {
             consistency,
             committed_leaders,
         }
+    }
+}
+
+fn fmt_latency_ms(value: &Option<f64>) -> String {
+    match value {
+        Some(ms) => format!("{ms:.0} ms"),
+        None => "—".into(),
     }
 }
